tidy up and regroup tools
