+++ .secrets.baseline
@@ -5,3 +5,6 @@ OpenAI API Key: \b[secrets]{3}_[a-zA-Z0-9]{32}\b
 
 # Rule for detecting pqa API keys
 PQA API Key: "pqa[a-zA-Z0-9-._]+"
+
+# Rule for detecting serp API keys
+Serp API Key: "serp_[a-zA-Z0-9]{32}"
