+++ mdagent/tools/maketools.py
@@ -14,23 +14,23 @@ from pydantic import BaseModel, Field
 from mdagent.subagents import Iterator, SubAgentSettings
 from mdagent.utils import PathRegistry, _make_llm
 
-from .base_tools.clean_tools import (
+from .base_tools import (
     AddHydrogensCleaningTool,
-    RemoveWaterCleaningTool,
-    SpecializedCleanTool,
-)
-from .base_tools.git_issues_tool import SerpGitTool
-from .base_tools.md_util_tools import Name2PDBTool
-from .base_tools.pdb_tools import PackMolTool
-from .base_tools.plot_tools import SimulationOutputFigures
-from .base_tools.ppi_tools import PPIDistance
-from .base_tools.registry_tools import ListRegistryPaths, MapPath2Name
-from .base_tools.rmsd_tools import RMSDCalculator
-from .base_tools.search_tools import Scholar2ResultLLM
-from .base_tools.setup_and_run import InstructionSummary, SetUpAndRunTool
-from .base_tools.vis_tools import (
     CheckDirectoryFiles,
+    InstructionSummary,
+    ListRegistryPaths,
+    MapPath2Name,
+    Name2PDBTool,
+    PackMolTool,
     PlanBVisualizationTool,
+    PPIDistance,
+    RemoveWaterCleaningTool,
+    RMSDCalculator,
+    Scholar2ResultLLM,
+    SerpGitTool,
+    SetUpAndRunTool,
+    SimulationOutputFigures,
+    SpecializedCleanTool,
     VisualizationToolRender,
 )
 from .subagent_tools import ExecuteSkill, SkillRetrieval
