+++ mypy.ini
@@ -14,19 +14,15 @@ python_version = 3.10
 # parts of our code. We should only add these settings for libraries that we're certain
 # don't have type hints or stub files.
 
-# Ignore missing imports from setuptools
-[mypy-setuptools.*]
+[mypy-ast.*]
+ignore_missing_imports = True
+
+ # Ignore missing imports from chromadb
+[mypy-Chroma.*]
 ignore_missing_imports = True
 
 # Ignore missing imports from dotenv
 [mypy-dotenv.*]
-ignore_missing_imports = True
-
-# ignore missing imports from matplotlib
-[mypy-matplotlib.*]
-ignore_missing_imports = True
-
-[mypy-ast.*]
 ignore_missing_imports = True
 
 # Ignore missing imports from langchain
@@ -35,11 +31,29 @@ ignore_missing_imports = True
 [mypy-langchain_experimental.*]
 ignore_missing_imports = True
 
-# Ignore missing imports from paperqa
-[mypy-paperqa.*]
+# ignore missing imports from matplotlib
+[mypy-matplotlib.*]
 ignore_missing_imports = True
 
-# Ignore missing imports from pqapi
+ # Ignore missing imports from mdanalysis
+[mypy-MDAnalysis.*]
+ignore_missing_imports = True
+
+# Ignore missing imports from nbformat
+[mypy-nbformat.*]
+ignore_missing_imports = True
+
+ # Ignore missing imports from numpy
+[mypy-numpy.*]
+ignore_missing_imports = True
+
+ # Ignore missing imports from openmm
+[mypy-openmm.*]
+ignore_missing_imports = True
+
+# Ignore missing imports from paperqa & pqapi
+[mypy-paperqa.*]
+ignore_missing_imports = True
 [mypy-pqapi.*]
 ignore_missing_imports = True
 
@@ -47,60 +61,44 @@ ignore_missing_imports = True
 [mypy-paperscraper.*]
 ignore_missing_imports = True
 
-# Ignore missing imports from pypdf.errors
-[mypy-pypdf.*]
-ignore_missing_imports = True
-
-# Ignore missing imports from rmrkl
-[mypy-rmrkl.*]
-ignore_missing_imports = True
-
-# Ignore missing imports from nbformat
-[mypy-nbformat.*]
-ignore_missing_imports = True
-
-# Ignore missing imports from requests
-[mypy-requests.*]
-ignore_missing_imports = True
-
  # Ignore missing imports from pdbfixer
 [mypy-pdbfixer.*]
 ignore_missing_imports = True
 
- # Ignore missing imports from openmm
-[mypy-openmm.*]
+# Ignore missing imports from pypdf.errors
+[mypy-pypdf.*]
 ignore_missing_imports = True
 
- # Ignore missing imports from numpy
-[mypy-numpy.*]
-ignore_missing_imports = True
-
- # Ignore missing imports from mdanalysis
-[mypy-MDAnalysis.*]
+#Ignore missing imports from pydantic
+[mypy-pydantic.*]
 ignore_missing_imports = True
 
  # Ignore missing imports from pytest
 [mypy-pytest.*]
 ignore_missing_imports = True
 
- # Ignore missing imports from chromadb
-[mypy-Chroma.*]
+# Ignore missing imports from requests
+[mypy-requests.*]
 ignore_missing_imports = True
 
- # Ignore missing imports from tiktoken
-[mypy-tiktoken.*]
+# Ignore missing imports from rmrkl
+[mypy-rmrkl.*]
 ignore_missing_imports = True
 
 # Ignore missing imports from serpapi
 [mypy-serpapi.*]
 ignore_missing_imports = True
 
+# Ignore missing imports from setuptools
+[mypy-setuptools.*]
+ignore_missing_imports = True
+
 #Ignore missing imports from subprocess
 [mypy-subprocess.*]
 ignore_missing_imports = True
 
-#Ignore missing imports from pydantic
-[mypy-pydantic.*]
+ # Ignore missing imports from tiktoken
+[mypy-tiktoken.*]
 ignore_missing_imports = True
 
 #Ignore missing imports from typing
