+++ setup.py
@@ -18,19 +18,19 @@ setup(
     packages=find_packages(),
     install_requires=[
         "paper-scraper @ git+https://github.com/blackadad/paper-scraper.git",
-        "langchain==0.0.336",
-        "langchain_experimental",
-        "paper-qa",
-        "pqapi",
-        "rmrkl",
-        "nbformat",
-        "requests",
-        "python-dotenv",
-        "matplotlib",
         "chromadb==0.3.29",
         "google-search-results",
-        "tiktoken",
+        "langchain==0.0.336",
+        "langchain_experimental",
+        "matplotlib",
+        "nbformat",
         "openai",
+        "paper-qa",
+        "python-dotenv",
+        "pqapi",
+        "requests",
+        "rmrkl",
+        "tiktoken",
     ],
     test_suite="tests",
     long_description=long_description,
